requirements.txt
requests==2.22.0
dateparser==1.0.0
Flask==1.1.2
flask_cors==3.0.10
flask_restful==0.3.8
flask_toastr==0.5.4
matplotlib==3.3.4
Pillow==8.1.1
PyJWT==2.0.1
pymysql==1.0.2
pytz==2021.1
secrets==1.0.2
Werkzeug==1.0.1
